requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.35.0
python-dotenv==1.0.0
requests>=2.31.0
supabase==2.18.1
openai==1.58.1
assemblyai==0.21.0
python-multipart==0.0.6
ffmpeg-python==0.2.0
pydub==0.25.1
moviepy==1.0.3
PyJWT>=2.10.1

